requirements.txt v0.2 closes #81 (#98)
* requirements.txt v0.1

Attempted to create a requirements.txt file.  From what I could read, it consists of the arguments in the commands for setup.sh (the package names).  Not entirely sure if any of it is right

* Update requirements.txt v0.2

I assume that we are using the most recent versions of these

* Update requirements.txt

Updated an issue I overlooked