SQL Fix Closes #104 (#121)
* Update requirements.txt

* Update 0.1?

Not really sure

* Update ErrorSQL.py

* Update main.py

* Update AuthenticationSQL.py

* Update AuthenticationSQL.py

Okay so I fixed how we get the stuffs from the SQL. There were a few things that were pretty screwy. So now we look up "authType" using "PiHostname" in the SQL, then look up the variable that authType is for ID in user. This is what is returned in machineAuth. For some reason, we had machineAuth returning the current machine, which makes no sense as we try to use it as a boolean later in the program but also as a machinetype in the SQL call in line 54, 55. This is like really stupid and needs to be fixed but I'm more concerned about fixing the algorithm than an event log. @gacosta, @gmross, and I will discuss a better way for us to do this. Also I'd like us to clear up a lot of the formatting for our code. It's very sporadic and all over the place where functions are. It'd be a lot of cleaning to do but I'd really like us to reorganize where we put functions. I found myself searching for a long time trying to find a specific function.

Anyways /rant. @ophilli does this fix the issue you had or do you have any further comments?

* Update AuthenticationSQL.py