PiHat Completion
References for footprints are placed and named more efficiently. The pi
header footprint is now on the back of the board and all connections to
it were corrected. Added "+"s to the AUX references to show which side
is positive and those reference texts are now what GPIO pin connects to
it. Some footprints that are for components that are polarized or need
to be placed a certain way were redone to show the orientation of the
component for more efficient placement.