Fixes Thor's Misery (#151)
* Update README.md

Added a period to the end of the sentence

* New Schematic for PiHat

* Finished PiHat v2 Design

* Neopixel Footprint Fix

Rearranged the neopixel terminals to correctly mount it. Moved around
silk screen references and changed the text to better understand the
design when soldering. Added the developers' names to the back of the
board.

* PiHat Completion

References for footprints are placed and named more efficiently. The pi
header footprint is now on the back of the board and all connections to
it were corrected. Added "+"s to the AUX references to show which side
is positive and those reference texts are now what GPIO pin connects to
it. Some footprints that are for components that are polarized or need
to be placed a certain way were redone to show the orientation of the
component for more efficient placement.

* Revert "PiHat Completion"

This reverts commit 6738365f23747a299ba35bc740daefcfdec640b0.

* Revert "Revert "PiHat Completion""

This reverts commit 9dfad16c77172f9cc3f0b0126633a712500f49b7.

* New piHat

* Merge remote-tracking branch 'refs/remotes/clemsonMakerspace/devel' into devel

# Conflicts:
#	Hardware/PiHat_v2/.pretty/Neo_RGB_LED.kicad_mod
#	Hardware/PiHat_v2/PiHat_v2.kicad_pcb
#	Hardware/PiHat_v2/PiHat_v2.kicad_pcb-bak
#	Hardware/PiHat_v2/PiHat_v2.net
#	Hardware/PiHat_v2/PiHat_v2.pro